requirements.txt
twitchAPI==2.1.0
Mastodon.py==1.5.1
tweepy==4.14.0
python-dotenv==1.0.0
atproto==0.0.33